ephys.TrialEvent and ActionEvent: change float to decimal(10, 5); reingest

+++ pipeline/ephys.py
@@ -198,7 +198,7 @@ class TrialEvent(dj.Imported):
     trial_event_id: smallint
     ---
     -> experiment.TrialEventType
-    trial_event_time : float  # (s) from session start (global time)
+    trial_event_time : Decimal(10, 5)  # (s) from session start (global time)
     """    
 
 
@@ -212,7 +212,7 @@ class ActionEvent(dj.Imported):
     action_event_id: smallint
     ---
     -> experiment.ActionEventType
-    action_event_time : float  # (s) from session start (global time)
+    action_event_time : Decimal(10, 5) # (s) from session start (global time)
     """
 
 @schema
